% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdf-read.R
\name{as.data.frame.bdf}
\alias{as.data.frame.bdf}
\title{Convert \code{bdf} font into a data.frame}
\usage{
\method{as.data.frame}{bdf}(x, ...)
}
\arguments{
\item{x}{\code{bdf} font object}

\item{...}{ignored}
}
\description{
Convert \code{bdf} font into a data.frame
}
