% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdf-read.R
\name{as.data.frame.bdf_char}
\alias{as.data.frame.bdf_char}
\title{Convert \code{bdf_char} object into a data.frame}
\usage{
\method{as.data.frame}{bdf_char}(x, ...)
}
\arguments{
\item{x}{\code{bdf_char} object}

\item{...}{ignored}
}
\description{
Convert \code{bdf_char} object into a data.frame
}
