% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{bdf_create_df}
\alias{bdf_create_df}
\title{Create a data.frame of the given string and font}
\usage{
bdf_create_df(bdf, text, line_height = NULL)
}
\arguments{
\item{bdf}{\code{bdf} font object as returned by \code{read_bdf()}}

\item{text}{string}

\item{line_height}{This value is used for vertical spacing between lines of text when
the text is wrapped.  If \code{line_height} is NULL, then use the pixel size
speficied in the font.  You may wish to set a value here if you to
ensure ascenders/descenders do not overlap on subsequent lines, or just
to space out the text a bit more.}
}
\value{
data.frame of x,y coordinates
}
\description{
Create a data.frame of the given string and font
}
