% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{bdf_extract_chars}
\alias{bdf_extract_chars}
\title{Get a list of \code{bdf_char} objects from a \code{bdf} font}
\usage{
bdf_extract_chars(bdf, text)
}
\arguments{
\item{bdf}{\code{bdf} font object as returned by \code{read_bdf()}}

\item{text}{string}
}
\value{
list of \code{bdf_char} objects. A space is used for any
        unknown characters, or the default character from the font if one is
        specified in the font header.
}
\description{
Get a list of \code{bdf_char} objects from a \code{bdf} font
}
