% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{bdf_print_char}
\alias{bdf_print_char}
\title{View a single \code{bdf_char}}
\usage{
bdf_print_char(bdf_char, zero = " ", one = "#")
}
\arguments{
\item{bdf_char}{single \code{bdf_char}}

\item{zero, one}{characters to use to represent zero and one}
}
\description{
View a single \code{bdf_char}
}
