% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{bdf_print_sample}
\alias{bdf_print_sample}
\title{Print sample text to screen}
\usage{
bdf_print_sample(
  bdf,
  text,
  width = NULL,
  wrap = TRUE,
  trim = TRUE,
  line_height = NULL,
  zero = " ",
  one = "#"
)
}
\arguments{
\item{bdf}{\code{bdf} font object as returned by \code{read_bdf()}}

\item{text}{string}

\item{width}{desired output width for output. Default: NULL means to
use the current console width}

\item{wrap}{wrap the text into separate lines using \code{base::strwrap()}.
Default: TRUE.   Note that \code{strwrap()} will only break lines
at whitespace, and resulting text may still overflow terminal width.
See \code{trim} argument.}

\item{trim}{trim the text at the consolve width. default: TRUE}

\item{line_height}{This value is used for vertical spacing between lines of text when
the text is wrapped.  If \code{line_height} is NULL, then use the pixel size
speficied in the font.  You may wish to set a value here if you to
ensure ascenders/descenders do not overlap on subsequent lines, or just
to space out the text a bit more.}

\item{zero, one}{characters to use to represent zero and one}
}
\value{
None
}
\description{
Print sample text to screen
}
\examples{
1 + 1

}
