% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-bdf.R
\name{bdfGrob}
\alias{bdfGrob}
\title{Convert a \code{bdf_str} object into grid grob object}
\usage{
bdfGrob(
  bdf,
  text,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  hjust = NULL,
  vjust = NULL,
  rot = 0,
  default.units = "points",
  name = NULL,
  gp = NULL,
  size = 2,
  shrink = 0.9,
  wrap = FALSE,
  width = 12,
  col = "black",
  fill = "black",
  alpha,
  lty,
  lwd,
  lex,
  lineend,
  linejoin,
  linemitre,
  fontsize,
  cex,
  fontfamily,
  fontface,
  lineheight
)
}
\arguments{
\item{bdf}{\code{bdf} font}

\item{text}{string}

\item{x, y, hjust, vjust, rot, default.units, name, gp, size, shrink}{other}

\item{wrap}{use \code{strwrap()} to wrap text into lines? Default: TRUE}

\item{width}{number of characters to use when wrap = TRUE. default: 12}

\item{col, fill, alpha, lty, lwd, lex, lineend, linejoin, linemitre, fontsize, cex, fontfamily, fontface, lineheight}{See documentation for \code{grid::gpar()}}
}
\value{
grid grob object
}
\description{
Convert a \code{bdf_str} object into grid grob object
}
\examples{
1 + 1
}
