% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\docType{data}
\name{bitmap_fonts}
\alias{bitmap_fonts}
\title{Bitmap font names}
\format{
An object of class \code{character} of length 9.
}
\usage{
bitmap_fonts
}
\description{
Bitmap font names
}
\keyword{datasets}
