% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\name{bitmap_text_matrix}
\alias{bitmap_text_matrix}
\title{Create a matrix of the given string and font}
\usage{
bitmap_text_matrix(text, font, line_height = NULL)
}
\arguments{
\item{text}{text}

\item{font}{bdf font name}

\item{line_height}{height}
}
\value{
matrix
}
\description{
Create a matrix of the given string and font
}
