% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\name{bitmap_text_raster}
\alias{bitmap_text_raster}
\title{Create a raster}
\usage{
bitmap_text_raster(text, font = "unifont", scale = 1, line_height = NULL)
}
\arguments{
\item{text}{text}

\item{font}{bitmap font name}

\item{scale}{Integer size scale factor. Default: 1.  Must be an integer value >= 1.
Scale up the matrix or raster result by this factor}

\item{line_height}{height}
}
\value{
raster
}
\description{
Create a raster
}
