% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdf.R
\name{create_coords}
\alias{create_coords}
\title{Create a data.frame of the given string and font}
\usage{
create_coords(text, font, line_height = NULL)
}
\arguments{
\item{text}{text}

\item{font}{bdf font name}

\item{line_height}{height}
}
\value{
data.frame of x,y coordinates
}
\description{
Create a data.frame of the given string and font
}
