% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a matrix of the given string and font}
\usage{
create_matrix(text, bdf, line_height = NULL)
}
\arguments{
\item{text}{text}

\item{bdf}{font}

\item{line_height}{height}
}
\value{
matrix
}
\description{
Create a matrix of the given string and font
}
