% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdf.R
\name{create_raster}
\alias{create_raster}
\title{Create a raster}
\usage{
create_raster(text, font, line_height = NULL)
}
\arguments{
\item{text}{text}

\item{font}{bdf font name}

\item{line_height}{height}
}
\value{
raster
}
\description{
Create a raster
}
