% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\docType{data}
\name{font_info}
\alias{font_info}
\title{Font information}
\format{
An object of class \code{list} of length 2.
}
\usage{
font_info
}
\description{
Currently on includes font names and unicode codepoints available within
each font
}
\keyword{datasets}
