% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bdf-text.R
\name{geom_text_bdf}
\alias{geom_text_bdf}
\title{Add text in a bitmap font to a ggplot}
\usage{
geom_text_bdf(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  nudge_x = 0,
  nudge_y = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, ..., nudge_x, nudge_y, na.rm, show.legend, inherit.aes}{See documentation for \code{ggplot2::geom_text()}}
}
\description{
Add text in a bitmap font to a ggplot
}
