% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{print.bdf}
\alias{print.bdf}
\title{S3 method for printing \code{bdf} font objects}
\usage{
\method{print}{bdf}(x, ...)
}
\arguments{
\item{x}{bdf font object as returned by \code{read_bdf()}}

\item{...}{ignored}
}
\value{
None
}
\description{
S3 method for printing \code{bdf} font objects
}
\examples{
# Print output from a font to the terminal
1 + 1
}
