% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{read_bdf_builtin}
\alias{read_bdf_builtin}
\title{Read one of the fonts included with this package}
\usage{
read_bdf_builtin(
  font_name = c("spleen-5x8.bdf", "spleen-6x12.bdf", "spleen-8x16.bdf",
    "spleen-12x24.bdf", "spleen-16x32.bdf", "spleen-32x64.bdf", "creep2-11.bdf",
    "cozette.bdf")
)
}
\arguments{
\item{font_name}{name of font}
}
\value{
bdf font
}
\description{
Read one of the fonts included with this package
}
\examples{
read_bdf_builtin()
}
