% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfont.R
\name{vector_text_coords_single_row}
\alias{vector_text_coords_single_row}
\title{Create data.frame of glyph information for the given line of text.}
\usage{
vector_text_coords_single_row(text, font, dx = 0)
}
\arguments{
\item{text}{single character string}

\item{font}{choose font: 'original' or 'smooth'. default: original}

\item{dx}{character spacing in original units}
}
\value{
data.frame with coordinates for all the glyphs with characters offset
       appropriately.  \code{char_idx} is the index of the character within
       the given text string
}
\description{
Create data.frame of glyph information for the given line of text.
}
