% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-text.R
\name{vector_text_matrix}
\alias{vector_text_matrix}
\title{Create a binary matrix rendering of the text}
\usage{
vector_text_matrix(text, font, scale = 10, dx = NULL, dy = NULL)
}
\arguments{
\item{text}{string}

\item{font}{fontname}

\item{scale}{scale factor for text rendering}

\item{dx}{extra spacing}

\item{dy}{extra spacing}
}
\value{
raster image
}
\description{
Create a binary matrix rendering of the text
}
\examples{
vector_text_matrix("Hello", font = "gridfont", scale = 1)
}
