% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-text.R
\name{vector_text_raster}
\alias{vector_text_raster}
\title{Create a raster rendering of the font}
\usage{
vector_text_raster(text, font, scale = 10, dx = NULL, dy = NULL)
}
\arguments{
\item{text}{string}

\item{font}{fontname}

\item{scale}{scale factor for text rendering}

\item{dx}{extra spacing}

\item{dy}{extra spacing}
}
\value{
raster image
}
\description{
Create a raster rendering of the font
}
\examples{
ras <- vector_text_raster("Hello", font = "gridfont", scale = 15)
plot(ras, interpolate = FALSE)
}
