% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap-text.R
\name{bitmap_text_coords}
\alias{bitmap_text_coords}
\title{Create a data.frame of the given string and font}
\usage{
bitmap_text_coords(text, font = "unifont", line_height = NULL)
}
\arguments{
\item{text}{text}

\item{font}{bitmap font name}

\item{line_height}{height}
}
\value{
data.frame of x,y coordinates
}
\description{
Create a data.frame of the given string and font
}
