% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-bdf.R
\name{print.bdf}
\alias{print.bdf}
\title{S3 method for printing \code{bdf} font objects}
\usage{
\method{print}{bdf}(x, text = "Handgloves", zero = " ", one = "#", ...)
}
\arguments{
\item{x}{bdf font object as returned by \code{read_bdf()}}

\item{text}{Example string to output to console in this font. Default: "Handgloves"}

\item{zero, one}{characters to use to represent zero and one}

\item{...}{ignored}
}
\description{
S3 method for printing \code{bdf} font objects
}
