% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-bdf.R
\name{read_bdf}
\alias{read_bdf}
\title{Read a BDF file}
\usage{
read_bdf(filename)
}
\arguments{
\item{filename}{path to BDF file}
}
\value{
list containing
\itemize{
\item{font_info - named list of global information about the font}
\item{font_info_raw - original representation of global information about the font}
\item{chars - list of \code{bdf_char} objects}
\item{idx - an index which maps from utf8 codepoint to position in \code{chars}}
}
}
\description{
Read a BDF file
}
