% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfont.R
\name{vector_text_coords}
\alias{vector_text_coords}
\title{Create data.frame of glyph information for the given text.}
\usage{
vector_text_coords(
  text,
  font = c("gridfont", "gridfont_smooth", "arcade"),
  dx = 0,
  dy = 0
)
}
\arguments{
\item{text}{single character string}

\item{font}{choose font: 'original' or 'smooth'. default: original}

\item{dx}{character spacing in original units}

\item{dy}{character spacing in original units}
}
\value{
data.frame with coordinates for all the glyphs with characters offset
       appropriately.  \code{char_idx} is the index of the character within
       the given text string
}
\description{
Text input can contain multiple lines separated by carriage returns
}
